Brilliant way of calculating Hamming's distance. Does not work

+++ BossPuzzle/Dao/FileFifteenPuzzleDao.cs
@@ -30,12 +30,12 @@ public class FileFifteenPuzzleDao : IDao<Board>
         var list = data[0].Split(' ');
         int columnSize = Parser.ToInt32(list[0]);
         int rowSize = Parser.ToInt32(list[1]);
-        var table = new long[columnSize][];
+        var table = new short[columnSize][];
         
         for (var i = 0; i < columnSize; i++)
         {
             var row = data[i + 1].Split(' ', StringSplitOptions.RemoveEmptyEntries);
-            table[i] = new long[rowSize];
+            table[i] = new short[rowSize];
 
             for (var j = 0; j < rowSize; j++)
             {
